Moved Fuse instance and results calculation to `SearchResults`
